requirements.txt
numpy==1.26.4
torch
torchvision
roma
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
tensorboard
pyglet<2
huggingface-hub[torch]>=0.22
viser
gradio
lpips
hydra-core
pillow==10.3.0
h5py
accelerate
transformers
scikit-learn
OpenEXR
pyrender==0.1.45
smplx
pyvista==0.42.3
spaces==0.19.4
git+https://github.com/mattloper/chumpy@9b045ff5d6588a24a0bab52c83f032e2ba433e17
